most tests pass after refactor
